added all dtos and changed decimal to float again as per the project description

+++ template/Cryptocop.Software.API.Models/Entities/OrderItem.cs
@@ -5,7 +5,7 @@ public class OrderItem
     public int Id { get; set; }
     public int OrderId { get; set; }
     public string ProductIdentifier { get; set; } = null!;
-    public decimal Quantity { get; set; }
-    public decimal UnitPrice { get; set; }
-    public decimal TotalPrice { get; set; }
+    public float Quantity { get; set; }
+    public float UnitPrice { get; set; }
+    public float TotalPrice { get; set; }
 }
